requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
openai==0.27.0
werkzeug==2.0.1
sqlalchemy==1.4.47